@FIR-1110: Updated code to support multiple models

+++ llm/server.go
@@ -1718,16 +1718,20 @@ func (s *llmServer) Close() error {
 
 	if s.cmd != nil {
 		slog.Debug("stopping llama server", "pid", s.Pid())
-		if err := s.cmd.Process.Kill(); err != nil {
-			return err
-		}
-		// if ProcessState is already populated, Wait already completed, no need to wait again
-		if s.cmd.ProcessState == nil {
-			slog.Debug("waiting for llama server to exit", "pid", s.Pid())
-			<-s.done
-		}
+		if true {
+			return nil
+		} else {
+			if err := s.cmd.Process.Kill(); err != nil {
+				return err
+			}
+			// if ProcessState is already populated, Wait already completed, no need to wait again
+			if s.cmd.ProcessState == nil {
+				slog.Debug("waiting for llama server to exit", "pid", s.Pid())
+				<-s.done
+			}
 
-		slog.Debug("llama server stopped", "pid", s.Pid())
+			slog.Debug("llama server stopped", "pid", s.Pid())
+		}
 	}
 
 	return nil
